requirements.txt
fastapi
uvicorn
Psycopg2-binary
